Implement contextChecker interface on TideContextPolicy
