+++ prow/tide/BUILD.bazel
@@ -2,10 +2,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "contextregister.go",
-        "tide.go",
-    ],
+    srcs = ["tide.go"],
     importpath = "k8s.io/test-infra/prow/tide",
     visibility = ["//visibility:public"],
     deps = [
@@ -36,10 +33,7 @@ filegroup(
 
 go_test(
     name = "go_default_test",
-    srcs = [
-        "contextregister_test.go",
-        "tide_test.go",
-    ],
+    srcs = ["tide_test.go"],
     embed = [":go_default_library"],
     deps = [
         "//prow/config:go_default_library",
